+++ src/themes/teams/components/Avatar/avatarStyles.ts
@@ -11,10 +11,10 @@ const avatarStyles: IComponentPartStylesInput<IAvatarPropsWithDefaults, any> = {
     height: pxToRem(size),
     width: pxToRem(size),
   }),
-  imageAvatar: (): ICSSInJSStyle => ({
+  image: (): ICSSInJSStyle => ({
     verticalAlign: 'top',
   }),
-  avatarNameContainer: ({ props: { size } }): ICSSInJSStyle => ({
+  label: ({ props: { size } }): ICSSInJSStyle => ({
     display: 'inline-block',
     width: pxToRem(size),
     height: pxToRem(size),
@@ -22,6 +22,7 @@ const avatarStyles: IComponentPartStylesInput<IAvatarPropsWithDefaults, any> = {
     fontSize: pxToRem(size / 2.333),
     verticalAlign: 'top',
     textAlign: 'center',
+    padding: '0px',
   }),
   status: ({ props, variables }): ICSSInJSStyle => ({
     position: 'absolute',